+++ config/splat.us.stwrp.yaml
@@ -34,8 +34,9 @@ segments:
       - [0x0, .data, header]
       - [0xB8, layers, rooms]
       - [0x1B8, .data, 6FD0]
-      - [0xF84, .data, entity_relic_orb]
-      - [0xFE8, .data, F420]
+      - [0xF84, .data, e_misc]
+      - [0xFE8, .data, e_particles]
+      - [0x1120, .data, e_room_fg]
       - [0x11AC, roomdef, g_Rooms]
       - [0x1228, layoutobj, D_80181228]
       - [0x1250, layoutobj, D_80181250]
@@ -74,26 +75,28 @@ segments:
       - [0x6E30, .rodata, 6FD0]
       - [0x6E70, .rodata, warp] # EntityWarpRoom
       - [0x6E90, .rodata, warp] # EntityWarpSmallRocks
-      - [0x6EA8, .rodata, 861C] # EntityRedDoor
+      - [0x6EA8, .rodata, e_red_door] # EntityRedDoor
       - [0x6EC0, rodata] # CollectGold
-      - [0x6F10, .rodata, D020]
-      - [0x6F30, .rodata, D020] # EntityEquipItemDrop
-      - [0x6F48, .rodata, entity_relic_orb]
-      - [0x6F80, .rodata, entity_message_box]
-      - [0x6F94, .rodata, F420] # EntityStageNamePopup
+      - [0x6F10, .rodata, st_common]
+      - [0x6F30, .rodata, e_collect] # EntityEquipItemDrop
+      - [0x6F48, .rodata, e_misc]
+      - [0x6F94, .rodata, e_stage_name] # EntityStageNamePopup
       - [0x6FC8, rodata]
       - [0x6FD0, c, st_debug]
       - [0x726C, c, e_breakable]
       - [0x73A0, c, warp]
-      - [0x81E8, c, random]
-      - [0x8218, c, update]
-      - [0x8514, c, update_stage_entities]
-      - [0x861C, c]
-      - [0xD020, c]
-      - [0xE2CC, c, entity_relic_orb]
-      - [0xED9C, c, entity_heart_drop]
-      - [0xEEC4, c, entity_message_box]
-      - [0xF420, c]
+      - [0x81E8, c, st_update]
+      - [0x861C, c, collision]
+      - [0x9DD8, c, create_entity]
+      - [0xA950, c, e_red_door]
+      - [0xB600, c, st_common]
+      - [0xCAB0, c, e_collect]
+      - [0xE01C, c, blit_char]
+      - [0xE2CC, c, e_misc]
+      - [0x117BC, c, e_stage_name]
+      - [0x12610, c, e_particles]
+      - [0x12E54, c, e_room_fg]
+      - [0x12F40, c, popup]
       - [0x13270, c, prim_helpers]
       - [0x13AA0, sbss] # bss?
   - [0x14768]
