+++ src/st/entity_red_door.h
@@ -1,5 +1,7 @@
 #include <game.h>
 
+#include "entity_is_near_player.h"
+
 extern u8 g_eRedDoorUV[3][8];
 extern u16 g_eRedDoorTiles[2][8];
 void EntityRedDoor(Entity* self) {
