+++ src/st/libstage.h
@@ -9,4 +9,3 @@
 #include "init_room_entities.h"
 #include "update_room_position.h"
 #include "create_entity_from_entity.h"
-#include "entity_is_near_player.h"
