+++ src/st/wrp/blit_char.c
@@ -1,0 +1,3 @@
+#include "wrp.h"
+
+#include "../blit_char.h"
