+++ src/st/wrp/collision.c
@@ -1,0 +1,5 @@
+#include "wrp.h"
+
+INCLUDE_ASM("st/wrp/nonmatchings/collision", TestCollisions);
+
+#include "../entity_damage_display.h"
