+++ src/st/wrp/create_entity.c
@@ -1,0 +1,3 @@
+#include "wrp.h"
+
+#include "../libstage.h"
