+++ src/st/wrp/e_particles.c
@@ -1,0 +1,115 @@
+#include "wrp.h"
+
+u16 D_801810A0[] = {
+    /* 10A0 */ 0x0820,
+    /* 10A2 */ 0x0840,
+    /* 10A4 */ 0x0860,
+    /* 10A6 */ 0x0880,
+    /* 10A8 */ 0x08C0,
+    /* 10AA */ 0x0900,
+    /* 10AC */ 0x0940,
+    /* 10AE */ 0x0A00,
+};
+
+u16 D_801810B0[] = {
+    0xFFFD, 0xFFFD, 0x0008, 0x0008, 0x00D0, 0x0068, 0x00D8, 0x0070,
+    0xFFFD, 0xFFFD, 0x0008, 0x0008, 0x00C8, 0x0068, 0x00D0, 0x0070,
+    0xFFF9, 0xFFF9, 0x0010, 0x000F, 0x00C8, 0x0070, 0x00D8, 0x007F,
+    0xFFF5, 0xFFF5, 0x0018, 0x0017, 0x0080, 0x0068, 0x0098, 0x007F,
+    0xFFF5, 0xFFF5, 0x0018, 0x0017, 0x0098, 0x0068, 0x00B0, 0x007F,
+    0xFFF5, 0xFFF5, 0x0018, 0x0017, 0x00B0, 0x0068, 0x00C8, 0x007F,
+};
+
+u32 D_80181110[] = {
+    /* 1110 */ 0x03030204,
+    /* 1114 */ 0x05030403,
+    /* 1118 */ 0x03030603,
+    /* 111C */ 0x00000000,
+};
+// The white flying orbs of energy that Alucard summons as part of the Soul
+// Steal spell
+void EntitySoulStealOrb(Entity* self) {
+    Primitive* prim;
+    s32 primIndex;
+    u16 *temp_d, temp_e;
+    s32 temp_a, temp_b;
+    u16 angle;
+
+    switch (self->step) {
+    case 0:
+        primIndex = g_api.AllocPrimitives(PRIM_GT4, 1);
+        if (primIndex == -1) {
+            DestroyEntity(self);
+            return;
+        }
+        InitializeEntity(g_InitializeData0);
+        g_PrimBuf[primIndex].drawMode = DRAW_HIDE;
+        self->primIndex = primIndex;
+        self->animSet = ANIMSET_DRA(0);
+        self->flags |= FLAG_HAS_PRIMS;
+        angle = func_8018C160(self, &PLAYER);
+        temp_a = self->posY.i.hi < 113;
+        temp_b = temp_a ^ 1;
+        if (self->posX.i.hi < PLAYER.posX.i.hi) {
+            temp_b = temp_a;
+        }
+        if (temp_b & 0xFFFF) {
+            self->ext.soulStealOrb.angle = angle - D_801810A0[Random() & 7];
+        } else {
+            angle += D_801810A0[Random() & 7];
+            self->ext.soulStealOrb.angle = angle;
+        }
+        self->ext.soulStealOrb.unk80 = 0x400;
+        self->ext.soulStealOrb.unk7E = 0;
+        self->hitboxState = 0;
+        break;
+
+    case 1:
+        self->ext.soulStealOrb.unk82++;
+        if (self->ext.soulStealOrb.unk82 == 16) {
+            self->hitboxState = 1;
+        }
+        if (self->hitFlags != 0) {
+            if (g_Player.unk56 == 0) {
+                g_Player.unk56 = 1;
+                g_Player.unk58 = 8;
+            }
+            DestroyEntity(self);
+            return;
+        }
+        if (self->rotX < 0x100) {
+            self->rotX = self->rotY += 0x10;
+        }
+        if (self->ext.soulStealOrb.unk7E < 0x200) {
+            self->ext.soulStealOrb.unk7E += 2;
+        }
+        if (self->ext.soulStealOrb.unk80 < 0x800) {
+            self->ext.soulStealOrb.unk80 += 4;
+        }
+        self->ext.soulStealOrb.angle = func_8018C1E0(
+            self->ext.soulStealOrb.unk7E, (u16)self->ext.soulStealOrb.angle,
+            0xffff & func_8018C160(self, &PLAYER));
+        UnkEntityFunc0(self->ext.soulStealOrb.angle & 0xFFFF,
+                       self->ext.soulStealOrb.unk80);
+        MoveEntity(self); // argument pass necessary to match
+        prim = &g_PrimBuf[self->primIndex];
+        AnimateEntity(&D_80181110, self);
+        angle = (float)(u32)self; // !FAKE
+        prim->tpage = 0x18;
+        prim->clut = 0x194;
+        temp_d = &D_801810B0[(u16)((8 * (u16)self->animCurFrame) - 8)];
+        prim->x0 = prim->x2 = self->posX.i.hi + *(temp_d++);
+        prim->y0 = prim->y1 = self->posY.i.hi + *(temp_d++);
+        prim->x1 = prim->x3 = prim->x0 + *(temp_d++);
+        prim->y2 = prim->y3 = prim->y0 + *(temp_d++);
+        prim->u0 = prim->u2 = *(temp_d++);
+        prim->v0 = prim->v1 = *(temp_d++);
+        prim->u1 = prim->u3 = *(temp_d++);
+        prim->v2 = prim->v3 = *(temp_d++);
+        prim->priority = self->zPriority;
+        prim->drawMode = DRAW_DEFAULT;
+        break;
+    }
+}
+
+#include "../entity_enemy_blood.h"
