+++ src/st/wrp/e_red_door.c
@@ -1,0 +1,3 @@
+#include "wrp.h"
+
+#include "../entity_red_door.h"
