+++ src/st/wrp/e_stage_name.c
@@ -1,0 +1,3 @@
+#include "wrp.h"
+
+#include "../entity_stage_name_popup.h"
