--- src/st/wrp/entity_heart_drop.c
@@ -1,35 +1,0 @@
-#include "wrp.h"
-
-void EntityHeartDrop(Entity* self) {
-    u16 temp_a0;
-    u16 temp_a0_2;
-    u16 var_a0;
-
-    if (self->step == 0) {
-        temp_a0 = self->params + 0x118;
-        self->ext.generic.unkB4 = temp_a0;
-        if ((D_8003BEEC[temp_a0 >> 3] >> (temp_a0 & 7)) & 1) {
-            DestroyEntity(self);
-            return;
-        }
-        temp_a0_2 = temp_a0 - 0x118;
-        var_a0 = ((u16*)D_80180608)[temp_a0_2];
-        if (var_a0 < 128) {
-            self->ext.generic.unkB8.unkFuncB8 = EntityPrizeDrop;
-        } else {
-            self->ext.generic.unkB8.unkFuncB8 = EntityEquipItemDrop;
-            var_a0 -= 128;
-        }
-        self->params = var_a0 + 0x8000;
-    } else {
-        temp_a0_2 = self->ext.generic.unkB4;
-        if (self->step < 5) {
-            if (self->hitFlags != 0) {
-                var_a0 = self->ext.generic.unkB4;
-                D_8003BEEC[temp_a0_2 >> 3] |= 1 << (var_a0 & 7);
-                self->step = 5;
-            }
-        }
-    }
-    self->ext.generic.unkB8.unkFuncB8(self);
-}
