--- src/st/wrp/entity_message_box.c
@@ -1,3 +1,0 @@
-#include "wrp.h"
-
-#include "../entity_message_box.h"
