--- src/st/wrp/entity_relic_orb.c
@@ -1,3 +1,0 @@
-#include "wrp.h"
-
-#include "../entity_relic_orb.h"
