--- src/st/wrp/random.c
@@ -1,3 +1,0 @@
-#include <game.h>
-
-#include "../random.h"
