+++ src/st/wrp/st_update.c
@@ -1,3 +1,7 @@
 #include <stage.h>
 
+#include "../random.h"
+
+#include "../update.h"
+
 #include "../update_stage_entities.h"
