--- src/st/wrp/update.c
@@ -1,3 +1,0 @@
-#include <game.h>
-
-#include "../update.h"
